Extract game loop and input handing into a package
